hti_web_api/requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==2.1
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
six==1.14.0
sqlparse==0.3.1
whitenoise==4.1.3
